+++ src/app.module.ts
@@ -1,29 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ProductsController } from './controllers/products.controller';
-import { CategoriesController } from './controllers/categories.controller';
-import { OrdersController } from './controllers/orders.controller';
-import { UsersController } from './controllers/users.controller';
-import { CustomersController } from './controllers/customers.controller';
-import { BrandsController } from './controllers/brands.controller';
-import { ProductsService } from './services/products.service';
-import { CategoriesService } from './services/categories.service';
-import { BrandsService } from './services/brands.service';
-import { UsersService } from './services/users.service';
-import { CustomersService } from './services/customers.service';
+import { UsersModule } from './users/users.module';
+import { ProductsModule } from './products/products.module';
 
 @Module({
-  imports: [],
-  controllers: [
-    AppController,
-    ProductsController,
-    CategoriesController,
-    OrdersController,
-    UsersController,
-    CustomersController,
-    BrandsController,
-  ],
-  providers: [AppService, ProductsService, CategoriesService, BrandsService, UsersService, CustomersService],
+  imports: [UsersModule, ProductsModule],
+  controllers: [AppController],
+  providers: [AppService],
 })
 export class AppModule {}
