+++ src/products/controllers/brands.controller.ts
@@ -8,8 +8,8 @@ import {
   Put,
 } from '@nestjs/common';
 import { ParseIntPipe } from 'src/common/parse-int.pipe';
-import { CreateBrandDto, UpdateBrandDto } from 'src/dtos/brand.dto';
-import { BrandsService } from 'src/services/brands.service';
+import { CreateBrandDto, UpdateBrandDto } from '../dtos/brand.dto';
+import { BrandsService } from '../services/brands.service';
 
 @Controller('brands')
 export class BrandsController {
