+++ src/products/controllers/products.controller.ts
@@ -11,8 +11,8 @@ import {
   HttpCode,
   // ParseIntPipe,
 } from '@nestjs/common';
-import { ProductsService } from 'src/services/products.service';
-import { ParseIntPipe } from './../common/parse-int.pipe';
+import { ProductsService } from 'src/products/services/products.service';
+import { ParseIntPipe } from './../../common/parse-int.pipe';
 import { CreateProductDto, UpdateProductDto } from './../dtos/products.dto';
 @Controller('products')
 export class ProductsController {
