+++ src/products/entities/product.entity.ts
