+++ src/products/products.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { ProductsController } from './controllers/products.controller';
+import { CategoriesController } from './controllers/categories.controller';
+import { BrandsController } from './controllers/brands.controller';
+import { ProductsService } from './services/products.service';
+import { CategoriesService } from './services/categories.service';
+import { BrandsService } from './services/brands.service';
+@Module({
+  controllers: [ProductsController, CategoriesController, BrandsController],
+  providers: [ProductsService, CategoriesService, BrandsService],
+})
+export class ProductsModule {}
